Split out HeartbeatConsumer into own class
