+++ src/main/java/com/mongodb/kafka/connect/source/heartbeat/HeartbeatManager.java
@@ -18,19 +18,12 @@ package com.mongodb.kafka.connect.source.heartbeat;
 
 import static com.mongodb.kafka.connect.source.MongoSourceTask.ID_FIELD;
 
-import java.time.Duration;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Properties;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import org.apache.kafka.clients.consumer.CommitFailedException;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.common.errors.WakeupException;
-import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.source.SourceRecord;
@@ -43,10 +36,6 @@ import com.mongodb.client.MongoChangeStreamCursor;
 
 public class HeartbeatManager implements AutoCloseable {
   private static final Logger LOGGER = LoggerFactory.getLogger(HeartbeatManager.class);
-  private static final String CONSUMER_GROUP_ID = "MONGODB_SOURCE_HEARTBEAT";
-  private static final String CONSUMER_DESERIALIZER =
-      "org.apache.kafka.common.serialization.BytesDeserializer";
-
   public static final String HEARTBEAT_KEY = "HEARTBEAT";
   private final Time time;
   private final MongoChangeStreamCursor<? extends BsonDocument> cursor;
@@ -75,7 +64,7 @@ public class HeartbeatManager implements AutoCloseable {
     this.canCreateHeartbeat = cursor != null && heartbeatIntervalMS > 0;
     this.heartbeatConsumer =
         canCreateHeartbeat && !bootstrapServers.isEmpty()
-            ? new HeartbeatConsumer(bootstrapServers)
+            ? new HeartbeatConsumer(bootstrapServers, heartbeatIntervalMS, heartbeatTopicName)
             : null;
   }
 
@@ -118,70 +107,6 @@ public class HeartbeatManager implements AutoCloseable {
   public void close() throws Exception {
     if (!isClosed.getAndSet(true) && heartbeatConsumer != null) {
       heartbeatConsumer.shutdown();
-    }
-  }
-
-  public class HeartbeatConsumer implements Runnable {
-    private final AtomicBoolean running;
-    private final KafkaConsumer<Bytes, Bytes> consumer;
-
-    public HeartbeatConsumer(final List<String> bootStrapServers) {
-      this.consumer = tryCreateConsumer(String.join(",", bootStrapServers)).orElse(null);
-      this.running = new AtomicBoolean(false);
-      if (consumer != null) {
-        new Thread(this).start();
-        LOGGER.info("Start heartbeat offset consumer");
-      }
-    }
-
-    public void run() {
-      running.set(true);
-      consumer.subscribe(Collections.singleton(heartbeatTopicName));
-      Duration pollDuration = Duration.ofMillis(heartbeatIntervalMS);
-      try {
-        while (running.get()) {
-          if (!consumer.poll(pollDuration).isEmpty()) {
-            try {
-              LOGGER.info("Syncing heartbeat offsets");
-              consumer.commitSync();
-            } catch (CommitFailedException e) {
-              // ignore any superseded commits by the connector
-            }
-          }
-        }
-      } catch (WakeupException e) {
-        // Ignore exception if closing
-        if (!running.get()) {
-          throw e;
-        }
-      } catch (Exception e) {
-        LOGGER.error("Heartbeat consumer exception", e);
-      } finally {
-        consumer.close();
-      }
-    }
-
-    public void shutdown() {
-      running.set(false);
-      if (consumer != null) {
-        consumer.wakeup();
-      }
-    }
-  }
-
-  private Optional<KafkaConsumer<Bytes, Bytes>> tryCreateConsumer(final String bootStrapServers) {
-    Properties props = new Properties();
-    props.setProperty("bootstrap.servers", bootStrapServers);
-    props.setProperty("group.id", CONSUMER_GROUP_ID);
-    props.setProperty("enable.auto.commit", "true");
-    props.setProperty("key.deserializer", CONSUMER_DESERIALIZER);
-    props.setProperty("value.deserializer", CONSUMER_DESERIALIZER);
-
-    try {
-      return Optional.of(new KafkaConsumer<>(props));
-    } catch (Exception e) {
-      LOGGER.error("Unable to create Heartbeat consumer", e);
-      return Optional.empty();
     }
   }
 }
